requirement.txt
Django==2.1.7
djangorestframework==3.9.1
get==2018.11.19
post==2018.11.20
public==2018.11.20
pytz==2018.9
query-string==2018.11.20
request==2018.11.20
